<lib> libnmath libarcc libdp * updated libnmath * updated libarcc * Adding initial support for libdp (device properties)

+++ src/lib/libnmath/nmath.c
@@ -1,12 +1,11 @@
-#include <stdint.h>
 #include <nmath.h>
 
-uint32_t clog2(uint32_t num)
+unsigned int clog2(unsigned int num)
 {
 	return (num > 1) ? (1 + clog2((num % 2) ? ((num + 1) / 2) : (num / 2))) : 0;
 }
 
-uint64_t ceiling(uint32_t num, uint32_t den)
+unsigned long ceiling(unsigned long num, unsigned long den)
 {
 	return (num % den) ? ((num / den) + 1) : (num / den);
 }
